// src/controllers/utasi.interface.ts
export interface ParapetInterface {
  parapet_id: number;
  id: string;
  parapet_status: string;
  parapet_size: string;
  station_name: string;
}

export interface BacklitInterface {
  backlit_id: number;
  id: string;
  backlit_status: string;
  station_name: string;
}

export interface StationInterface {
  station_id: number;
  station_name: string;
}